+++ TiendaVirtualOrtiz/Controllers/UsuarioController.cs
@@ -31,12 +31,22 @@ namespace TiendaVirtualOrtiz.Controllers
 
         public IActionResult Create()
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             return View();
         }
 
         [HttpPost]
         public IActionResult Create(Usuario usuario)
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             _context.Usuarios.Add(usuario);
             _context.SaveChanges();
 
@@ -46,6 +56,11 @@ namespace TiendaVirtualOrtiz.Controllers
         //Formulario editar
         public IActionResult Edit(int id)
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             var usuario = _context.Usuarios.Find(id);
             ViewBag.Usuarios = _context.Usuarios.ToList();
 
@@ -57,6 +72,11 @@ namespace TiendaVirtualOrtiz.Controllers
         [HttpPost]
         public IActionResult Edit(Usuario usuario)
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             _context.Usuarios.Update(usuario);
             _context.SaveChanges();
 
@@ -66,6 +86,17 @@ namespace TiendaVirtualOrtiz.Controllers
         //Eliminar usuario
         public IActionResult Delete(int id)
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            var rol = HttpContext.Session.GetString("Rol");//Solo admin puede eliminar
+            if (rol != "admin")
+            {
+                return RedirectToAction("Index");
+            }
+
             var usuario = _context.Usuarios.Find(id);
 
             _context.Usuarios.Remove(usuario);
